Restructure and renaming.
