--- seclea_ai/exceptions.py
@@ -1,6 +1,0 @@
-class AuthenticationError(Exception):
-    """
-    Raised for any authentication error.
-    """
-
-    pass
