--- seclea_ai/svc/api/errors.py
@@ -1,12 +1,0 @@
-class APIError(Exception):
-    pass
-
-
-def throws_api_err(f):
-    def call(*args, **kwargs):
-        try:
-            return f(*args, **kwargs)
-        except Exception as e:
-            raise APIError(e)
-
-    return call
